refactor code

+++ src/main/java/site/viosmash/english/dto/response/AudioResponse.java
@@ -1,6 +1,7 @@
 package site.viosmash.english.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,8 +14,8 @@ public class AudioResponse {
     private int id;
     private int duration;
     private String format;
-    private int sampleRate;
-    private int fileSize;
+    private double sampleRate;
+    private double fileSize;
     private String fileUrl;
     private int pagesId;
 }
